Move most of main into common
