+++ registry/common.go
@@ -1,0 +1,66 @@
+package registry
+
+import (
+	"encoding/json"
+	"errors"
+	"log"
+	"net/http"
+	"os"
+)
+
+// healthHandler is a http.handler that returns the version
+type healthHandler struct {
+	healthInfo *map[string]string
+}
+
+// ServeHTTP implements http.Handler
+func (h *healthHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("content-type", "application/json")
+	if r.Method == http.MethodGet && r.URL.Path == "/health" {
+		jsonBytes, err := json.Marshal(*h.healthInfo)
+		if err != nil {
+			log.Println("ERROR:", err)
+			InternalServerError(w, r, err)
+			return
+		}
+		w.WriteHeader(http.StatusOK)
+		_, errw := w.Write(jsonBytes)
+		if errw != nil {
+			log.Println("ERROR:", errw)
+		}
+	} else {
+		NotFound(w, r)
+		return
+	}
+}
+
+func getAuthToken() (string, error) {
+	authToken, found := os.LookupEnv("AUTH_TOKEN")
+	if !found {
+		return "", errors.New("AUTH_TOKEN not found, set it to a secret token or '' to disable authentication")
+	}
+	return authToken, nil
+}
+
+// The main entrypoint for the service
+func Run(registryH IRegistryClient, healthInfo *map[string]string, listen string) {
+	authToken, err := getAuthToken()
+	if err != nil {
+		log.Fatalln(err)
+	}
+
+	health := healthHandler{
+		healthInfo: healthInfo,
+	}
+
+	mux := http.NewServeMux()
+	mux.Handle("/health", &health)
+
+	CreateServer(mux, registryH, authToken)
+
+	log.Printf("Listening on %v\n", listen)
+	errw := http.ListenAndServe(listen, mux)
+	if errw != nil {
+		log.Fatalln(errw)
+	}
+}
