+++ registry/main_test.go
@@ -1,4 +1,4 @@
-package main
+package registry
 
 import (
 	"fmt"
